+++ website-learning/src/adminPage/MgtSyllabus.jsx
@@ -2,12 +2,19 @@ import React, { useState, useEffect } from "react";
 import Modal from "react-bootstrap/Modal";
 import { useDispatch } from "react-redux";
 import { Form, Button, FormLabel } from "react-bootstrap";
+import Table from "@mui/material/Table";
+import TableBody from "@mui/material/TableBody";
+import TableCell from "@mui/material/TableCell";
+import TableHead from "@mui/material/TableHead";
+import TableRow from "@mui/material/TableRow";
 import {
   showsyllabus,
   insertsyllabus,
   editsyllabus,
   deletesyllabus,
 } from "../slice/DataSlice";
+import { Link } from "react-router-dom";
+import { saveToLocalStorage } from "../LocalStorage/localstorage";
 
 function MgtSyllabus() {
   const dispatch = useDispatch();
@@ -17,6 +24,17 @@ function MgtSyllabus() {
   const [showEdit, setshowEdit] = useState(false);
   const [datamodal, setDatamodal] = useState([]);
   const [update, setupdate] = useState({ sylla_name: "" });
+
+  const loadData = () => {
+    dispatch(showsyllabus())
+      .then((result) => {
+        setshowdata(result.payload);
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  };
+
   const AddClose = () => {
     setShowAdd(false);
     loadData();
@@ -35,6 +53,10 @@ function MgtSyllabus() {
   };
   //   on click insert value
   const onInsrt = () => {
+    if (!insert.sylla_name) {
+      alert("กรุณาป้อนข้อมูลให้ครบก่อนบันทึก!!");
+      return;
+    }
     let body = {
       sylla_name: insert.sylla_name,
     };
@@ -42,10 +64,11 @@ function MgtSyllabus() {
     dispatch(insertsyllabus(body))
       .then((result) => {
         AddClose();
-        setShowAdd({
+        setinsert({
           sylla_name: "",
         });
         loadData();
+        setShowAdd(false);
         // show success notification
         // notify();
       })
@@ -53,22 +76,7 @@ function MgtSyllabus() {
         console.log(err);
       });
   };
-  useEffect(() => {
-    loadData();
-  }, []);
-  const loadData = () => {
-    dispatch(showsyllabus())
-      .then((result) => {
-        setshowdata(result.payload);
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  };
-  // reload
-  useEffect(() => {
-    loadData();
-  }, []);
+
   const EditClose = () => {
     setshowEdit(false);
   };
@@ -108,37 +116,67 @@ function MgtSyllabus() {
   };
   //   Delete
   const onDelete = (id) => {
-    dispatch(deletesyllabus(id))
-      .then((result) => {
-        if (result.payload && result.payload.error) {
-          console.log(result.payload.error);
-        } else {
-          loadData();
-        }
-      })
-      .catch((err) => {
-        console.log(err);
-      });
+    if (window.confirm("Are you sure you want to delete?")) {
+      dispatch(deletesyllabus(id))
+        .then((result) => {
+          if (result.payload && result.payload.error) {
+            console.log(result.payload.error);
+          } else {
+            loadData();
+          }
+        })
+        .catch((err) => {
+          console.log(err);
+        });
+    }
   };
+
+  // reload
+  useEffect(() => {
+    loadData();
+  }, []);
+
+  saveToLocalStorage("syllaId", null);
+  const onClick = (id) => {
+    // console.log(id);
+    saveToLocalStorage("syllaId", id);
+  };
+
   return (
     <>
       <Button className="button" variant="primary" onClick={AddShow}>
         ADD
       </Button>
-      <table>
-        <thead>
-          <tr>
-            <th>ชื่อหลักสูตร</th>
-            <th>confix</th>
-          </tr>
-        </thead>
-        <tbody>
+      <Table stickyHeader aria-label="sticky table">
+        <TableHead className="TableHead">
+          <TableRow>
+            <TableCell>
+              <p className="headerC">ชื่อหลักสูตร</p>
+            </TableCell>
+            <TableCell>
+              <p className="headerC">จัดการวิชา</p>
+            </TableCell>
+            <TableCell>
+              <p className="headerC">confix</p>
+            </TableCell>
+          </TableRow>
+        </TableHead>
+        <TableBody>
           {showdata.map((data) => {
             const { sylla_id, sylla_name } = data;
             return (
-              <tr key={sylla_id}>
-                <td>{sylla_name}</td>
-                <td>
+              <TableRow key={sylla_id}>
+                <TableCell>{sylla_name}</TableCell>
+                <TableCell>
+                  <Link
+                    className="linkshow"
+                    to={"/admin/mgtSubject"}
+                    onClick={() => onClick(sylla_id)}
+                  >
+                    เพิ่มวิชา
+                  </Link>
+                </TableCell>
+                <TableCell>
                   <Button
                     variant="btn btn-secondary"
                     onClick={() => EditShow(data)}
@@ -153,12 +191,12 @@ function MgtSyllabus() {
                   >
                     DELETE
                   </Button>
-                </td>
-              </tr>
+                </TableCell>
+              </TableRow>
             );
           })}
-        </tbody>
-      </table>
+        </TableBody>
+      </Table>
       {/* Insert Data */}
       <Modal show={showAdd} onHide={AddClose}>
         <Modal.Header closeButton>
